requirements.txt
aiohttp>=3.8.0
async-timeout==3.0.1
attrs==20.3.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
discord.py>=2.4.0
dnspython==2.3.0
email-validator==1.3.1
firebase-admin==5.3.0
google-api-core==2.11.0
google-api-python-client==2.82.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
# grpcio==1.51.3
grpcio-status==1.51.3
httplib2==0.22.0
idna==3.1
msgpack==1.0.5
# multidict==5.1.0
notion-client==2.2.1
oauthlib==3.2.2
fuzzywuzzy==0.18.0
# python-Levenshtein==0.21.1
proto-plus==1.22.2
protobuf==4.22.1
pyairtable==1.0.0.post1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.15.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.15